requirements.txt
Django==3.1
djangorestframework==3.12.2
django-test-without-migrations==0.6
django-cors-headers==3.6.0
Faker==5.6.1
